SYN-120: Changed drug ndc's datatype from int to long.

+++ syncura360_spring/src/main/java/com/syncura360/dto/Drug/DrugDeletionDTO.java
@@ -7,5 +7,5 @@ import jakarta.validation.constraints.NotNull;
 @Getter
 public class DrugDeletionDTO {
     @NotNull(message = "NDC is required.")
-    private Integer ndc;
+    private Long ndc;
 }
